+++ test/integration/tests/button/props.js
@@ -5,7 +5,7 @@ import { wrapPromise } from 'belter/src';
 import { FUNDING } from '@paypal/sdk-constants/src';
 import { ZalgoPromise } from 'zalgo-promise/src';
 
-import { createTestContainer, destroyTestContainer, IPHONE6_USER_AGENT, mockProp } from '../common';
+import { createTestContainer, destroyTestContainer, IPHONE6_USER_AGENT } from '../common';
 
 describe(`paypal button component props`, () => {
 
@@ -15,112 +15,6 @@ describe(`paypal button component props`, () => {
 
     afterEach(() => {
         destroyTestContainer();
-    });
-
-    it('should not render a paylater button when the experiment disables it', () => {
-
-        return ZalgoPromise.try(() => {
-            return wrapPromise(({ expect, avoid }) => {
-                const mockEligibility = mockProp(window.__TEST_FUNDING_ELIGIBILITY__, 'paylater', {
-                    eligible: true,
-                    products: {
-                        paylater: {
-                            eligible: true,
-                            variant:  'experimentable'
-                        },
-                        payIn4: {
-                            eligible: false,
-                            variant:  'experimentable'
-                        }
-                    }
-                });
-
-                window.localStorage.setItem('disable_paylater', true);
-
-                let onRender = ({ xprops }) => {
-                    const renderedButtons = xprops.renderedButtons;
-                    const experiment = xprops.experiment;
-
-                    if (!experiment.disablePaylater) {
-                        throw new Error(`disablePaylater should be true. ${  JSON.stringify(experiment) }`);
-                    }
-
-                    if (JSON.stringify(renderedButtons).includes('paylater')) {
-                        throw new Error(`paylater button is being rendered but it should not be. ${ renderedButtons }`);
-                    }
-                };
-
-                const instance = window.paypal.Buttons({
-                    test: {
-                        action:   'checkout',
-                        onRender: (...args) => onRender(...args)
-                    },
-
-
-                    onApprove: avoid('onApprove'),
-                    onCancel:  avoid('onCancel')
-
-                });
-
-
-                if (instance.isEligible()) {
-                    onRender = expect('onRender', onRender);
-                    return instance.render('#testContainer').then(() => mockEligibility.cancel());
-                }
-            });
-        });
-    });
-
-    it('should render a paylater button when the experiment does not disable it', () => {
-
-        return ZalgoPromise.try(() => {
-            return wrapPromise(({ expect, avoid }) => {
-                const mockEligibility = mockProp(window.__TEST_FUNDING_ELIGIBILITY__, 'paylater', {
-                    eligible: true,
-                    products: {
-                        paylater: {
-                            eligible: true,
-                            variant:  ''
-                        },
-                        payIn4: {
-                            eligible: false,
-                            variant:  ''
-                        }
-                    }
-                });
-
-                let onRender = ({ xprops }) => {
-                    const renderedButtons = xprops.renderedButtons;
-                    const experiment = xprops.experiment;
-
-                    if (experiment.disablePaylater) {
-                        throw new Error(`disablePaylater should be falsy ${ JSON.stringify(experiment) }`);
-                    }
-
-                    if (!JSON.stringify(renderedButtons).includes('paylater')) {
-                        throw new Error(`paylater button is not being rendered but it should be. ${ renderedButtons }`);
-                    }
-                };
-
-                const instance = window.paypal.Buttons({
-                    test: {
-                        action:   'checkout',
-                        onRender: (...args) => onRender(...args)
-                    },
-
-
-                    onApprove: avoid('onApprove'),
-                    onCancel:  avoid('onCancel')
-
-                });
-
-
-                if (instance.isEligible()) {
-                    onRender = expect('onRender', onRender);
-                    return instance.render('#testContainer').then(() => mockEligibility.cancel());
-                }
-            });
-        });
     });
 
     it('should render an Apple Pay button if applePaySupport is true', () => {
@@ -272,62 +166,6 @@ describe(`paypal button component props`, () => {
                 if (instance.isEligible()) {
                     onRender = expect('onRender', onRender);
                     return instance.render('#testContainer');
-                }
-            });
-        });
-    });
-
-    it('should render a paylater button when the experiment when standalone integration', () => {
-
-        return ZalgoPromise.try(() => {
-            return wrapPromise(({ expect, avoid }) => {
-                const mockEligibility = mockProp(window.__TEST_FUNDING_ELIGIBILITY__, 'paylater', {
-                    eligible: true,
-                    products: {
-                        paylater: {
-                            eligible: true,
-                            variant:  'experimentable'
-                        },
-                        payIn4: {
-                            eligible: false,
-                            variant:  'experimentable'
-                        }
-                    }
-                });
-
-                window.localStorage.setItem('disable_paylater', true);
-
-                let onRender = ({ xprops }) => {
-                    const renderedButtons = xprops.renderedButtons;
-                    const experiment = xprops.experiment;
-
-                    if (experiment.disablePaylater) {
-                        throw new Error(`disablePaylater should not exist ${ JSON.stringify(experiment) }`);
-                    }
-
-                    if (!JSON.stringify(renderedButtons).includes('paylater')) {
-                        throw new Error(`paylater button is not being rendered but it should be. ${ renderedButtons }`);
-                    }
-                };
-
-                const instance = window.paypal.Buttons({
-                    test: {
-                        action:   'checkout',
-                        onRender: (...args) => onRender(...args)
-                    },
-
-                    fundingSource: 'paylater',
-
-
-                    onApprove: avoid('onApprove'),
-                    onCancel:  avoid('onCancel')
-
-                });
-
-
-                if (instance.isEligible()) {
-                    onRender = expect('onRender', onRender);
-                    return instance.render('#testContainer').then(() => mockEligibility.cancel());
                 }
             });
         });
